Refactor + move impl package inside core
