+++ pinot-core/src/main/java/org/apache/pinot/core/segment/processing/lifecycle/impl/SegmentDeletionEventDetails.java
@@ -1,0 +1,48 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.pinot.core.segment.processing.lifecycle.impl;
+
+import java.util.List;
+import org.apache.pinot.core.segment.processing.lifecycle.SegmentLifecycleEventDetails;
+import org.apache.pinot.core.segment.processing.lifecycle.SegmentLifecycleEventType;
+
+public class SegmentDeletionEventDetails implements SegmentLifecycleEventDetails {
+  private final List<String> _segmentsDeleted;
+  private final String _tableName;
+
+  public SegmentDeletionEventDetails(String tableName, List<String> segmentsDeleted) {
+    _tableName = tableName;
+    _segmentsDeleted = segmentsDeleted;
+  }
+
+  @Override
+  public SegmentLifecycleEventType getType() {
+    return SegmentLifecycleEventType.DELETION;
+  }
+
+  @Override
+  public List<String> getSegments() {
+    return _segmentsDeleted;
+  }
+
+  @Override
+  public String getTableNameWithType() {
+    return _tableName;
+  }
+}
